Create cartpole3.py
i still havent figured everything out yet but based on running this multiple times, I've learned:
by initializing observation of the cart (observation = env.reset() ) we have a set of data that represents the location of the cart and the pole (1x4); I do not know which each of the 4 sets of data correspond to. action (action = 0 if np.matmul(observation,parameters) < 0 else 1) is applied, by taking the the data set from observations and is then multiplied by the parameters which allows for the actions to alternate between 1 and 0 with matrix multiplication -- if the matmul yields a number less than 0 then action = 0 else action = 1 and then is fed into the action space.
I was reading up on the cartpole program and read something about using the reward(total reward) as a weight to help guide the algorithm and also used to supervise it as well.
more reading and studying required!